requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
numpy==2.4.3
pandas==3.0.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
pymssql==2.3.13
pyodbc==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
SQLAlchemy==2.0.48
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
Werkzeug==3.1.7
